Reorganize LeftMenu files for future PRs
