+++ src/NexusMods.App.UI/LeftMenu/ILeftMenuViewModel.cs
@@ -1,5 +1,6 @@
 using System.Collections.ObjectModel;
 using NexusMods.Abstractions.UI;
+using NexusMods.App.UI.LeftMenu.Items;
 using NexusMods.App.UI.WorkspaceSystem;
 
 namespace NexusMods.App.UI.LeftMenu;
