+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/ApplyControlDesignViewModel.cs
