+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/ApplyControlView.axaml
