+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/ApplyControlViewModel.cs
