+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/IApplyControlViewModel.cs
