+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/LaunchButtonDesignViewModel.cs
