+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/LaunchButtonView.axaml.cs
