+++ src/NexusMods.App.UI/LeftMenu/Items/ApplyControl/LaunchButtonViewModel.cs
