+++ src/NexusMods.App.UI/LeftMenu/Items/ILeftMenuItemViewModel.cs
@@ -1,6 +1,6 @@
 using NexusMods.Abstractions.UI;
 
-namespace NexusMods.App.UI.LeftMenu;
+namespace NexusMods.App.UI.LeftMenu.Items;
 
 public interface ILeftMenuItemViewModel : IViewModelInterface
 {
