+++ src/NexusMods.App.UI/LeftMenu/Items/LeftMenuCollectionItemViewModel.cs
@@ -1,9 +1,8 @@
 using NexusMods.Abstractions.Loadouts;
-using NexusMods.App.UI.LeftMenu.Items;
 
-namespace NexusMods.App.UI.LeftMenu.Loadout;
+namespace NexusMods.App.UI.LeftMenu.Items;
 
-public class LeftMenuCollectionViewModel : IconViewModel
+public class LeftMenuCollectionItemViewModel : IconViewModel
 {
     /// <summary>
     /// The group id of the collection.
