+++ src/NexusMods.App.UI/NexusMods.App.UI.csproj
@@ -391,19 +391,10 @@
         <Compile Update="LeftMenu\Loadout\LoadoutLeftMenuDesignViewModel.cs">
           <DependentUpon>ILoadoutLeftMenuViewModel.cs</DependentUpon>
         </Compile>
-        <Compile Update="LeftMenu\Items\LaunchButtonDesignViewModel.cs">
-          <DependentUpon>ILaunchButtonViewModel.cs</DependentUpon>
-        </Compile>
-        <Compile Update="LeftMenu\Items\LaunchButtonViewModel.cs">
-          <DependentUpon>ILaunchButtonViewModel.cs</DependentUpon>
-        </Compile>
         <Compile Update="LeftMenu\Items\IconViewModel.cs">
           <DependentUpon>IIconViewModel.cs</DependentUpon>
         </Compile>
         <Compile Update="LeftMenu\Items\IApplyControlDesignViewModel.cs">
-          <DependentUpon>IApplyControlViewModel.cs</DependentUpon>
-        </Compile>
-        <Compile Update="LeftMenu\Items\ApplyControlViewModel.cs">
           <DependentUpon>IApplyControlViewModel.cs</DependentUpon>
         </Compile>
         <Compile Update="Controls\ModInfo\Dummy\ViewModFilesView.axaml.cs">
@@ -450,9 +441,6 @@
         </Compile>
         <Compile Update="Controls\Trees\ModFileTreeViewModel.cs">
           <DependentUpon>IFileTreeViewModel.cs</DependentUpon>
-        </Compile>
-        <Compile Update="LeftMenu\Items\ApplyControlDesignViewModel.cs">
-          <DependentUpon>IApplyControlViewModel.cs</DependentUpon>
         </Compile>
         <Compile Update="Pages\MyGames\MyGamesDesignViewModel.cs">
           <DependentUpon>IMyGamesViewModel.cs</DependentUpon>
@@ -707,6 +695,23 @@
         </Compile>
         <Compile Update="Pages\Diagnostics\List\DiagnosticListDesignViewModel.cs.old">
           <DependentUpon>IDiagnosticListViewModel.cs</DependentUpon>
+        </Compile>
+        <Compile Update="LeftMenu\Items\IconDesignViewModel.cs">
+          <DependentUpon>IIconViewModel.cs</DependentUpon>
+        </Compile>
+        <Compile Update="LeftMenu\Items\ApplyControl\ApplyControlView.axaml.cs">
+          <DependentUpon>ApplyControlView.axaml</DependentUpon>
+          <SubType>Code</SubType>
+        </Compile>
+        <Compile Update="LeftMenu\Items\ApplyControl\LaunchButtonView.axaml.cs">
+          <DependentUpon>LaunchButtonView.axaml</DependentUpon>
+          <SubType>Code</SubType>
+        </Compile>
+        <Compile Update="LeftMenu\Items\LeftMenuCollectionItemViewModel.cs">
+          <DependentUpon>IconViewModel.cs</DependentUpon>
+        </Compile>
+        <Compile Update="LeftMenu\Items\IIconViewModel.cs">
+          <DependentUpon>ILeftMenuItemViewModel.cs</DependentUpon>
         </Compile>
     </ItemGroup>
